rework logs to include step/job breakdown
